Feature/41 prepare to publish (#50)
* added cookies support

* move i18n support to plugins folder

* add footer with
- language selector
- motto
- Terms of service (link)

* set default collection name

* set user locale after login

* remove foter messages
moved to global

* update translation names, also
update route for terms of service

* add terms of service route

* add terms of service view

* rename literals

* add terms of service

* add services logos

* show services logos

* update terms texts

* remove obsolete logos

* remove uneeded manifest.json file

* add meta description tag

* add meta keywords tag

* remove variant attribute from buttons

* remove sendgrid from footer

* remove language duplicated literals

* update version

* update production url

* remove unused 'loading' wheel

* show feedback while loading collection names

* add og attributes
