+++ index.html
@@ -6,8 +6,19 @@
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <link rel="icon" href="/favicon.ico">
     <title>WikiHooku</title>
-    <meta name="description" content="Wiki from hawaian Wikiwiki - Quick & Hooku from hawaian Hookuku - Compare. WikiHooku aims to compare data from diferents entities in the Wikipedia. First one is Lifetime & Age of Persons">
+    <meta name="description" content="Wikiwiki: Quick & Hookuku: Compare. WikiHooku compares data from differents entities in the Wikipedia.">
     <meta name="keywords" content="wikipedia compare persons age lifetime"/>
+
+    <meta property="og:type" content="website">
+    <meta property="og:title" content="WikiHooku compares data from Wikipedia">
+    <meta property="og:description" content="Wikiwiki: Quick & Hookuku: Compare. WikiHooku compares data from differents entities in the Wikipedia.">
+    <meta property="og:site_name" content="WikiHooku">
+    <meta property="og:url" content="https://www.wikihooku.org/">
+    <meta property="og:image" content="https://www.wikihooku.org/img/icons/android-chrome-512x512.png">
+    <meta property="og:image:width" content="512">
+    <meta property="og:image:height" content="512">
+    <meta property="og:image:type" content="image/png">
+
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700,900">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@mdi/font@latest/css/materialdesignicons.min.css">
   </head>
